// lib/api/admin.ts
export interface UserUpdateData {
  firstName?: string
  lastName?: string
  email?: string
  phone?: string
  role?: string
  userType?: string
  isVerified?: boolean
  isActive?: boolean
  isTrainee?: boolean
  autofillEnabled?: boolean
  operatorAirtableId?: string
  organizationId?: string | null
}
// ...
export interface UserCreateData {
  firstName: string
  lastName: string
  email?: string
  phone?: string
  role: string
  userType?: string
  organizationId?: string | null
  isVerified?: boolean
}